Returns a function to send emails using the emayili package.

This returns a function that can be used with the login_server(). Specifically, the function takes two parameters, to_email and message.

Usage

emayili_emailer(
  email_host = NULL,
  email_port = NULL,
  email_username = NULL,
  email_password = NULL,
  from_email = NULL
)

Arguments

email_host

SMTP email host.

email_port

SMPT email port.

email_username

username for the SMTP server.

email_password

password for the SMTP server.

from_email

the from email address sent from login_server().

Value

returns a function to send an email using the emayili package.

See also